# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Review API",
        "description": "This is an open and free to use API for developers to test their client projects.\r\n        All API calls require you to pass in <br> your API Key with the querystring\r\n        <b>?apikey=########-####-####-####-############</b>, <br><br>Please\r\n        register for your own API Key today at https://reviews.simpleapi.dev/register ",
        "version": "1.0"
    },
    "paths": {
...
        "/api/reviews": {
            "get": {
                "tags": [
                    "Review"
                ],
                "summary": "Retrieve all reviews",
                "operationId": "review.index",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "All api calls require the ?apikey querystring",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "spam",
                        "in": "query",
                        "description": "?spam=true to get reviews marked as spam",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "bad",
                        "in": "query",
                        "description": "?bad=true to get reviews marked as inappropriate/bad",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "helpful",
                        "in": "query",
                        "description": "?helpful=true to get reviews that are helpful",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A collection of reviews"
                    }
                }
            },
            "post": {
                "tags": [
                    "Review"
                ],
                "summary": "Create a new review.  There's a limit of 100 reviews per developer account.",
                "operationId": "review.post",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "All api calls require the ?apikey querystring",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Submit review field values",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "slug"
                                ],
                                "properties": {
                                    "slug": {
                                        "type": "string"
                                    },
                                    "rating": {
                                        "type": "number"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "body": {
                                        "type": "string"
                                    },
                                    "bad": {
                                        "type": "boolean"
                                    },
                                    "helpful_counter": {
                                        "type": "integer"
                                    },
                                    "unhelpful_counter": {
                                        "type": "integer"
                                    },
                                    "related_slug": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "author_email": {
                                        "type": "string"
                                    },
                                    "author_slug": {
                                        "type": "string"
                                    },
                                    "meta": {
                                        "type": "object"
                                    },
                                    "datetime": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Review created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "slug": {
                                            "type": "string"
                                        },
                                        "rating": {
                                            "type": "number"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "body": {
                                            "type": "string"
                                        },
                                        "bad": {
                                            "type": "boolean"
                                        },
                                        "helpful_counter": {
                                            "type": "integer"
                                        },
                                        "unhelpful_counter": {
                                            "type": "integer"
                                        },
                                        "related_slug": {
                                            "type": "string"
                                        },
                                        "author": {
                                            "type": "string"
                                        },
                                        "author_email": {
                                            "type": "string"
                                        },
                                        "author_slug": {
                                            "type": "string"
                                        },
                                        "meta": {
                                            "type": "object"
                                        },
                                        "datetime": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/authors/{author_slug}/reviews": {
            "get": {
                "tags": [
                    "Review"
                ],
                "summary": "Retrieve all reviews by author",
                "operationId": "review.indexByUser",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "All api calls require the ?apikey querystring",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "author_slug",
                        "in": "path",
                        "description": "The slug of the author",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "spam",
                        "in": "query",
                        "description": "?spam=true to get reviews marked as spam",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "bad",
                        "in": "query",
                        "description": "?bad=true to get reviews marked as inappropriate/bad",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "helpful",
                        "in": "query",
                        "description": "?helpful=true to get reviews that are helpful",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A collection of reviews"
                    }
                }
            }
        },
        "/api/related/{related_slug}/reviews": {
            "get": {
                "tags": [
                    "Review"
                ],
                "summary": "Retrieve all reviews having a related slug",
                "operationId": "review.indexByRelated",
                "parameters": [
                    {
                        "name": "apikey",
                        "in": "query",
                        "description": "All api calls require the ?apikey querystring",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "related_slug",
                        "in": "path",
                        "description": "The slug of the related grouping.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "spam",
                        "in": "query",
                        "description": "?spam=true to get reviews marked as spam",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "bad",
                        "in": "query",
                        "description": "?bad=true to get reviews marked as inappropriate/bad",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "helpful",
                        "in": "query",
                        "description": "?helpful=true to get reviews that are helpful",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A collection of reviews"
                    }
                }
            }
        },
...
    }
}